there

+++ voidstar/size2str.cc
@@ -5,7 +5,7 @@
 /// 991337 --> "991,337"
 std::string size2str(size_t size) {
   const auto digits = std::ceil(std::log10(size));
-  const auto nonits = static_cast<int>(std::ceil(digits / 3) -1);
+  const auto nonits = static_cast<size_t>(std::ceil(digits / 3) -1);
   std::string str(static_cast<size_t>(digits)+nonits, '*');
 
   int written = 0;
